ateom-microvm: drop OnDemand restore and the merge path it required
